add scanning dw & voa top-story

+++ cleanup.sh
@@ -2,6 +2,7 @@
 #
 
 pids=`cat .pid/*.pid`
-echo kill -9 $pids
-rm .pid/*.pid
-
+if [ ! -z "$pids" ]; then
+  echo kill -9 $pids
+  rm .pid/*.pid
+fi
